pyproject.toml
[project]
name = "kad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    'pandas',
    "google-api-python-client",
    "streamlit",
    "pytest",
    "python-dotenv",
    "openai>=1.97.1",
    "xlsxwriter>=3.2.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
